+++ internal/primitive/container.go
@@ -14,12 +14,14 @@ import (
 	"github.com/turbot/pipe-fittings/schema"
 )
 
-type Container struct{}
+type Container struct {
+	FullyQualifiedStepName string
+}
 
 var containerCache = map[string]*container.Container{}
 var containerCacheMutex sync.Mutex
 
-func (e *Container) ValidateInput(ctx context.Context, i modconfig.Input) error {
+func (cp *Container) ValidateInput(ctx context.Context, i modconfig.Input) error {
 
 	// Validate the name attribute
 	if i[schema.LabelName] == nil {
@@ -184,39 +186,41 @@ func convertMapToStrings(input map[string]interface{}) map[string]string {
 	return result
 }
 
-func (e *Container) Run(ctx context.Context, input modconfig.Input) (*modconfig.Output, error) {
-	if err := e.ValidateInput(ctx, input); err != nil {
+func (cp *Container) Run(ctx context.Context, input modconfig.Input) (*modconfig.Output, error) {
+	if err := cp.ValidateInput(ctx, input); err != nil {
 		return nil, err
 	}
 
-	c, err := e.getFromCacheOrNew(ctx, input)
+	c, err := cp.getFromCacheOrNew(ctx, input, cp.FullyQualifiedStepName)
 	if err != nil {
 		return nil, err
 	}
 
+	cConfig := container.ContainerRunConfig{}
+
 	if input[schema.AttributeTypeCmd] != nil {
-		c.Cmd = convertToSliceOfString(input[schema.AttributeTypeCmd].([]interface{}))
+		cConfig.Cmd = convertToSliceOfString(input[schema.AttributeTypeCmd].([]interface{}))
 	}
 
 	if input[schema.AttributeTypeEnv] != nil {
-		c.Env = convertMapToStrings(input[schema.AttributeTypeEnv].(map[string]interface{}))
+		cConfig.Env = convertMapToStrings(input[schema.AttributeTypeEnv].(map[string]interface{}))
 	}
 
 	if input[schema.AttributeTypeEntryPoint] != nil {
-		c.EntryPoint = convertToSliceOfString(input[schema.AttributeTypeEntryPoint].([]interface{}))
+		cConfig.EntryPoint = convertToSliceOfString(input[schema.AttributeTypeEntryPoint].([]interface{}))
 	}
 
 	if input[schema.AttributeTypeUser] != nil {
-		c.User = input[schema.AttributeTypeUser].(string)
+		cConfig.User = input[schema.AttributeTypeUser].(string)
 	}
 
 	if input[schema.AttributeTypeWorkdir] != nil {
-		c.Workdir = input[schema.AttributeTypeWorkdir].(string)
+		cConfig.Workdir = input[schema.AttributeTypeWorkdir].(string)
 	}
 
 	if input[schema.AttributeTypeReadOnly] != nil {
 		readOnly := input[schema.AttributeTypeReadOnly].(bool)
-		c.ReadOnly = &readOnly
+		cConfig.ReadOnly = &readOnly
 	}
 
 	if input[schema.AttributeTypeTimeout] != nil {
@@ -233,7 +237,7 @@ func (e *Container) Run(ctx context.Context, input modconfig.Input) (*modconfig.
 
 		// Convert milliseconds to seconds, and round up to the nearest second
 		timeoutInSeconds := int64(math.Ceil(float64(timeoutInMs) / 1000))
-		c.Timeout = &timeoutInSeconds
+		cConfig.Timeout = &timeoutInSeconds
 	}
 
 	if input[schema.AttributeTypeCpuShares] != nil {
@@ -246,7 +250,7 @@ func (e *Container) Run(ctx context.Context, input modconfig.Input) (*modconfig.
 		default:
 			break
 		}
-		c.CpuShares = &cpuShares
+		cConfig.CpuShares = &cpuShares
 	}
 
 	if input[schema.AttributeTypeMemory] != nil {
@@ -259,7 +263,7 @@ func (e *Container) Run(ctx context.Context, input modconfig.Input) (*modconfig.
 		default:
 			break
 		}
-		c.Memory = &memory
+		cConfig.Memory = &memory
 	}
 
 	if input[schema.AttributeTypeMemoryReservation] != nil {
@@ -272,7 +276,7 @@ func (e *Container) Run(ctx context.Context, input modconfig.Input) (*modconfig.
 		default:
 			break
 		}
-		c.MemoryReservation = &memoryReservation
+		cConfig.MemoryReservation = &memoryReservation
 	}
 
 	if input[schema.AttributeTypeMemorySwap] != nil {
@@ -285,7 +289,7 @@ func (e *Container) Run(ctx context.Context, input modconfig.Input) (*modconfig.
 		default:
 			break
 		}
-		c.MemorySwap = &memorySwap
+		cConfig.MemorySwap = &memorySwap
 	}
 
 	if input[schema.AttributeTypeMemorySwappiness] != nil {
@@ -298,7 +302,7 @@ func (e *Container) Run(ctx context.Context, input modconfig.Input) (*modconfig.
 		default:
 			break
 		}
-		c.MemorySwappiness = &memorySwappiness
+		cConfig.MemorySwappiness = &memorySwappiness
 	}
 
 	// Construct the output
@@ -306,7 +310,7 @@ func (e *Container) Run(ctx context.Context, input modconfig.Input) (*modconfig.
 		Data: map[string]interface{}{},
 	}
 
-	containerID, exitCode, err := c.Run()
+	containerID, exitCode, err := c.Run(cConfig)
 	if err != nil {
 		if e, ok := err.(perr.ErrorModel); !ok {
 			output.Errors = []modconfig.StepError{
@@ -339,9 +343,8 @@ func (e *Container) Run(ctx context.Context, input modconfig.Input) (*modconfig.
 	return &output, nil
 }
 
-func (e *Container) getFromCacheOrNew(ctx context.Context, input modconfig.Input) (*container.Container, error) {
-	name := input[schema.LabelName].(string)
-	c := containerCache[name]
+func (cp *Container) getFromCacheOrNew(ctx context.Context, input modconfig.Input, stepFullName string) (*container.Container, error) {
+	c := containerCache[stepFullName]
 
 	// if Dockerfile source path changed ignore cache & rebuild
 	if c != nil && input[schema.AttributeTypeSource] != nil && c.Source != input[schema.AttributeTypeSource].(string) {
@@ -364,7 +367,7 @@ func (e *Container) getFromCacheOrNew(ctx context.Context, input modconfig.Input
 		container.WithContext(context.Background()),
 		container.WithRunContext(ctx),
 		container.WithDockerClient(docker.GlobalDockerClient),
-		container.WithName(name),
+		container.WithName(stepFullName),
 	)
 	if err != nil {
 		return nil, perr.InternalWithMessage("Error creating container config with the provided options:" + err.Error())
@@ -383,7 +386,7 @@ func (e *Container) getFromCacheOrNew(ctx context.Context, input modconfig.Input
 		return nil, perr.InternalWithMessage("failed loading container config: " + err.Error())
 	}
 
-	containerCache[name] = c
+	containerCache[stepFullName] = c
 
 	return c, nil
 }